Move parse_raw_data to raw_data_parsers
